Update README.md
I used cbind function to create the data frame so Frequency, BP, First, Second, and FinalD ( as FinalDecision) define the columns and the patient's number are the rows. 
We also have a missing value in First dataset and I used "na.rm=T" argument to exclude the missing value from analysis in boxplot.